moved Imran's code a bit to fit the react structure better
