+++ client/src/components/TrainSearch/TrainSearch.js
@@ -3,13 +3,13 @@ import React, { Component } from 'react'
 // import components
 import Dropdown from './Dropdown/Dropdown'
 import Button from './Button/Button'
-import DetailsView from '../AppRouter/Views/DetailsView'
+import TrainDetails from './TrainDetails/TrainDetails'
 
 class TrainSearch extends Component {
     constructor(props) {
         super(props)
         this.state = {
-            select: "destination",
+            select: undefined,
             data: this.props.data,
             buttonVal: undefined,
         }
@@ -18,7 +18,6 @@ class TrainSearch extends Component {
     // select change handler
     handleSelectChange = (e) => {
         this.setState({ select: e.target.value })
-        //console.log(this.state.select);
     }
 
     // input change handler
@@ -38,7 +37,9 @@ class TrainSearch extends Component {
                     onClick={this.handleOnCLick} 
                 />
                 <h2>Train Details</h2>
-                <DetailsView trainNum={this.state.buttonVal} />
+                <TrainDetails 
+                    trainNum={this.state.buttonVal} 
+                />
             </div>
         )
     }
